requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
pandas==2.2.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
langchain==0.1.6
langchain-openai==0.0.5
tiktoken==0.5.2
chromadb==0.4.22
plotly==5.18.0
folium==0.15.1
streamlit-folium==0.16.0
pinecone-client==3.0.0

